gatra-agent/pyproject.toml
[project]
name = "gatra-agent"
version = "0.1.0"
description = "GATRA LangGraph agent backend for CopilotKit"
requires-python = ">=3.12,<3.13"
dependencies = [
    "langgraph>=0.3.25,<1.1.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-groq>=0.2.0",
    "copilotkit>=0.1.84",
    "ag-ui-langgraph>=0.0.29",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "pyjwt>=2.12.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.30",
    "ruff>=0.5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[dependency-groups]
dev = [
    "python-docx>=1.2.0",
]
